Elevate docstring/comment processing into base, and split filed
