+++ tests/utils/test_describe_model.py
@@ -58,6 +58,7 @@ class TestBasic(test.TestCase):
                 "table": "straightfields",
                 "abstract": False,
                 "description": "Straight auto-mapped fields",
+                "docstring": None,
                 "unique_together": [["chars", "blip"]],
                 "pk_field": {
                     "name": "eyedee",
@@ -71,6 +72,7 @@ class TestBasic(test.TestCase):
                     "indexed": True,
                     "default": None,
                     "description": "Da PK",
+                    "docstring": None,
                 },
                 "data_fields": [
                     {
@@ -85,6 +87,7 @@ class TestBasic(test.TestCase):
                         "indexed": True,
                         "default": None,
                         "description": "Some chars",
+                        "docstring": None,
                     },
                     {
                         "name": "blip",
@@ -98,6 +101,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": "BLIP",
                         "description": None,
+                        "docstring": None,
                     },
                     {
                         "name": "fk_id",
@@ -111,12 +115,14 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     },
                     {
                         "db_column": "o2o_id",
                         "db_field_types": {"": "INT"},
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                         "field_type": "IntField",
                         "generated": False,
                         "indexed": True,
@@ -138,6 +144,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     }
                 ],
                 "backward_fk_fields": [
@@ -151,12 +158,14 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     }
                 ],
                 "o2o_fields": [
                     {
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                         "field_type": "OneToOneFieldInstance",
                         "generated": False,
                         "indexed": True,
@@ -171,6 +180,7 @@ class TestBasic(test.TestCase):
                     {
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                         "field_type": "BackwardOneToOneRelation",
                         "generated": False,
                         "indexed": False,
@@ -191,6 +201,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "M2M to myself",
+                        "docstring": None,
                     },
                     {
                         "name": "rel_from",
@@ -202,6 +213,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "M2M to myself",
+                        "docstring": None,
                     },
                 ],
             },
@@ -218,6 +230,7 @@ class TestBasic(test.TestCase):
                 "table": "straightfields",
                 "abstract": False,
                 "description": "Straight auto-mapped fields",
+                "docstring": None,
                 "unique_together": [["chars", "blip"]],
                 "pk_field": {
                     "name": "eyedee",
@@ -231,6 +244,7 @@ class TestBasic(test.TestCase):
                     "indexed": True,
                     "default": None,
                     "description": "Da PK",
+                    "docstring": None,
                 },
                 "data_fields": [
                     {
@@ -245,6 +259,7 @@ class TestBasic(test.TestCase):
                         "indexed": True,
                         "default": None,
                         "description": "Some chars",
+                        "docstring": None,
                     },
                     {
                         "name": "blip",
@@ -258,6 +273,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": "BLIP",
                         "description": None,
+                        "docstring": None,
                     },
                     {
                         "name": "fk_id",
@@ -271,6 +287,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     },
                     {
                         "name": "o2o_id",
@@ -284,6 +301,7 @@ class TestBasic(test.TestCase):
                         "indexed": True,
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                     },
                 ],
                 "fk_fields": [
@@ -298,6 +316,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     }
                 ],
                 "backward_fk_fields": [
@@ -311,12 +330,14 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     }
                 ],
                 "o2o_fields": [
                     {
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                         "field_type": OneToOneFieldInstance,
                         "generated": False,
                         "indexed": True,
@@ -331,6 +352,7 @@ class TestBasic(test.TestCase):
                     {
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                         "field_type": fields.BackwardOneToOneRelation,
                         "generated": False,
                         "indexed": False,
@@ -351,6 +373,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "M2M to myself",
+                        "docstring": None,
                     },
                     {
                         "name": "rel_from",
@@ -362,6 +385,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "M2M to myself",
+                        "docstring": None,
                     },
                 ],
             },
@@ -378,6 +402,7 @@ class TestBasic(test.TestCase):
                 "table": "sometable",
                 "abstract": False,
                 "description": "Source mapped fields",
+                "docstring": "A Docstring.",
                 "unique_together": [["chars", "blip"]],
                 "pk_field": {
                     "name": "eyedee",
@@ -391,6 +416,7 @@ class TestBasic(test.TestCase):
                     "indexed": True,
                     "default": None,
                     "description": "Da PK",
+                    "docstring": None,
                 },
                 "data_fields": [
                     {
@@ -405,6 +431,7 @@ class TestBasic(test.TestCase):
                         "indexed": True,
                         "default": None,
                         "description": "Some chars",
+                        "docstring": None,
                     },
                     {
                         "name": "blip",
@@ -418,6 +445,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": "BLIP",
                         "description": None,
+                        "docstring": "A docstring comment",
                     },
                     {
                         "name": "fk_id",
@@ -431,6 +459,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     },
                     {
                         "name": "o2o_id",
@@ -444,6 +473,7 @@ class TestBasic(test.TestCase):
                         "indexed": True,
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                     },
                 ],
                 "fk_fields": [
@@ -458,6 +488,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     }
                 ],
                 "backward_fk_fields": [
@@ -471,12 +502,14 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     }
                 ],
                 "o2o_fields": [
                     {
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                         "field_type": "OneToOneFieldInstance",
                         "generated": False,
                         "indexed": True,
@@ -491,6 +524,7 @@ class TestBasic(test.TestCase):
                     {
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                         "field_type": "BackwardOneToOneRelation",
                         "generated": False,
                         "indexed": False,
@@ -511,6 +545,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "M2M to myself",
+                        "docstring": None,
                     },
                     {
                         "name": "rel_from",
@@ -522,6 +557,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "M2M to myself",
+                        "docstring": None,
                     },
                 ],
             },
@@ -538,6 +574,7 @@ class TestBasic(test.TestCase):
                 "table": "sometable",
                 "abstract": False,
                 "description": "Source mapped fields",
+                "docstring": "A Docstring.",
                 "unique_together": [["chars", "blip"]],
                 "pk_field": {
                     "name": "eyedee",
@@ -551,6 +588,7 @@ class TestBasic(test.TestCase):
                     "indexed": True,
                     "default": None,
                     "description": "Da PK",
+                    "docstring": None,
                 },
                 "data_fields": [
                     {
@@ -565,6 +603,7 @@ class TestBasic(test.TestCase):
                         "indexed": True,
                         "default": None,
                         "description": "Some chars",
+                        "docstring": None,
                     },
                     {
                         "name": "blip",
@@ -578,6 +617,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": "BLIP",
                         "description": None,
+                        "docstring": "A docstring comment",
                     },
                     {
                         "name": "fk_id",
@@ -591,6 +631,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     },
                     {
                         "name": "o2o_id",
@@ -604,6 +645,7 @@ class TestBasic(test.TestCase):
                         "indexed": True,
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                     },
                 ],
                 "fk_fields": [
@@ -618,6 +660,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     }
                 ],
                 "backward_fk_fields": [
@@ -631,12 +674,14 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "Tree!",
+                        "docstring": None,
                     }
                 ],
                 "o2o_fields": [
                     {
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                         "field_type": OneToOneFieldInstance,
                         "generated": False,
                         "indexed": True,
@@ -651,6 +696,7 @@ class TestBasic(test.TestCase):
                     {
                         "default": None,
                         "description": "Line",
+                        "docstring": None,
                         "field_type": fields.BackwardOneToOneRelation,
                         "generated": False,
                         "indexed": False,
@@ -671,6 +717,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "M2M to myself",
+                        "docstring": None,
                     },
                     {
                         "name": "rel_from",
@@ -682,6 +729,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": "M2M to myself",
+                        "docstring": None,
                     },
                 ],
             },
@@ -698,6 +746,7 @@ class TestBasic(test.TestCase):
                 "table": "uuidpkmodel",
                 "abstract": False,
                 "description": None,
+                "docstring": None,
                 "unique_together": [],
                 "pk_field": {
                     "name": "id",
@@ -711,6 +760,7 @@ class TestBasic(test.TestCase):
                     "indexed": True,
                     "default": "<function uuid.uuid4>",
                     "description": None,
+                    "docstring": None,
                 },
                 "data_fields": [],
                 "fk_fields": [],
@@ -725,6 +775,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": None,
+                        "docstring": None,
                     },
                 ],
                 "o2o_fields": [],
@@ -740,6 +791,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": None,
+                        "docstring": None,
                     }
                 ],
             },
@@ -756,6 +808,7 @@ class TestBasic(test.TestCase):
                 "table": "uuidpkmodel",
                 "abstract": False,
                 "description": None,
+                "docstring": None,
                 "unique_together": [],
                 "pk_field": {
                     "name": "id",
@@ -769,6 +822,7 @@ class TestBasic(test.TestCase):
                     "indexed": True,
                     "default": uuid.uuid4,
                     "description": None,
+                    "docstring": None,
                 },
                 "data_fields": [],
                 "fk_fields": [],
@@ -783,6 +837,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": None,
+                        "docstring": None,
                     },
                 ],
                 "o2o_fields": [],
@@ -798,6 +853,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": None,
+                        "docstring": None,
                     }
                 ],
             },
@@ -819,6 +875,7 @@ class TestBasic(test.TestCase):
                         "db_field_types": {"": "VARCHAR(50)"},
                         "default": None,
                         "description": None,
+                        "docstring": None,
                         "field_type": "CharField",
                         "generated": False,
                         "indexed": False,
@@ -832,6 +889,7 @@ class TestBasic(test.TestCase):
                         "db_field_types": {"": "CHAR(36)", "postgres": "UUID"},
                         "default": None,
                         "description": None,
+                        "docstring": None,
                         "field_type": "UUIDField",
                         "generated": False,
                         "indexed": False,
@@ -845,6 +903,7 @@ class TestBasic(test.TestCase):
                         "db_field_types": {"": "CHAR(36)", "postgres": "UUID"},
                         "default": None,
                         "description": None,
+                        "docstring": None,
                         "field_type": "UUIDField",
                         "generated": False,
                         "indexed": True,
@@ -855,10 +914,12 @@ class TestBasic(test.TestCase):
                     },
                 ],
                 "description": None,
+                "docstring": None,
                 "fk_fields": [
                     {
                         "default": None,
                         "description": None,
+                        "docstring": None,
                         "field_type": "ForeignKeyFieldInstance",
                         "generated": False,
                         "indexed": False,
@@ -875,6 +936,7 @@ class TestBasic(test.TestCase):
                     {
                         "default": None,
                         "description": None,
+                        "docstring": None,
                         "field_type": "OneToOneFieldInstance",
                         "generated": False,
                         "indexed": True,
@@ -890,6 +952,7 @@ class TestBasic(test.TestCase):
                     "db_field_types": {"": "CHAR(36)", "postgres": "UUID"},
                     "default": "<function uuid.uuid4>",
                     "description": None,
+                    "docstring": None,
                     "field_type": "UUIDField",
                     "generated": False,
                     "indexed": True,
@@ -913,7 +976,8 @@ class TestBasic(test.TestCase):
                 "app": "models",
                 "table": "jsonfields",
                 "abstract": False,
-                "description": None,
+                "description": "This model contains many JSON blobs",
+                "docstring": "This model contains many JSON blobs",
                 "unique_together": [],
                 "pk_field": {
                     "name": "id",
@@ -927,6 +991,7 @@ class TestBasic(test.TestCase):
                     "indexed": True,
                     "default": None,
                     "description": None,
+                    "docstring": None,
                 },
                 "data_fields": [
                     {
@@ -941,6 +1006,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": None,
+                        "docstring": None,
                     },
                     {
                         "name": "data_null",
@@ -954,6 +1020,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": None,
+                        "docstring": None,
                     },
                     {
                         "name": "data_default",
@@ -967,6 +1034,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": "{'a': 1}",
                         "description": None,
+                        "docstring": None,
                     },
                 ],
                 "fk_fields": [],
@@ -987,7 +1055,8 @@ class TestBasic(test.TestCase):
                 "app": "models",
                 "table": "jsonfields",
                 "abstract": False,
-                "description": None,
+                "description": "This model contains many JSON blobs",
+                "docstring": "This model contains many JSON blobs",
                 "unique_together": [],
                 "pk_field": {
                     "name": "id",
@@ -1001,6 +1070,7 @@ class TestBasic(test.TestCase):
                     "indexed": True,
                     "default": None,
                     "description": None,
+                    "docstring": None,
                 },
                 "data_fields": [
                     {
@@ -1015,6 +1085,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": None,
+                        "docstring": None,
                     },
                     {
                         "name": "data_null",
@@ -1028,6 +1099,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": None,
                         "description": None,
+                        "docstring": None,
                     },
                     {
                         "name": "data_default",
@@ -1041,6 +1113,7 @@ class TestBasic(test.TestCase):
                         "indexed": False,
                         "default": {"a": 1},
                         "description": None,
+                        "docstring": None,
                     },
                 ],
                 "fk_fields": [],
