+++ tortoise/contrib/pydantic/base.py
@@ -1,0 +1,77 @@
+from typing import TYPE_CHECKING, List, Type
+
+import pydantic
+from pydantic import BaseModel  # pylint: disable=E0611
+
+from tortoise import fields
+
+if TYPE_CHECKING:  # pragma: nocoverage
+    from tortoise.models import Model
+    from tortoise.queryset import QuerySet
+
+
+def _get_fetch_fields(
+    pydantic_class: "Type[PydanticModel]", model_class: "Type[Model]"
+) -> List[str]:
+    """
+    Recursively collect fields needed to fetch
+    :param pydantic_class: The pydantic model class
+    :param model_class: The tortoise model class
+    :return: The list of fields to be fetched
+    """
+    fetch_fields = []
+    for field_name, field_type in pydantic_class.__annotations__.items():
+        origin = getattr(field_type, "__origin__", None)
+        if origin in (list, List):
+            field_type = field_type.__args__[0]
+        # noinspection PyProtectedMember
+        if field_name in model_class._meta.fetch_fields and issubclass(field_type, PydanticModel):
+            subclass_fetch_fields = _get_fetch_fields(
+                field_type, getattr(field_type.__config__, "orig_model")
+            )
+            if subclass_fetch_fields:
+                fetch_fields.extend([field_name + "__" + f for f in subclass_fetch_fields])
+            else:
+                fetch_fields.append(field_name)
+    return fetch_fields
+
+
+class PydanticModel(BaseModel):
+    """ Custom Pydantic BaseModel for Tortoise objects """
+
+    class Config:
+        orm_mode = True  # It should be in ORM mode to convert tortoise data to pydantic
+
+    # noinspection PyMethodParameters
+    @pydantic.validator("*", pre=True, each_item=False)  # It is a classmethod!
+    def _tortoise_convert(cls, value):  # pylint: disable=E0213
+        # Computed fields
+        if callable(value):
+            return value()
+        # Convert ManyToManyRelation to list
+        elif isinstance(value, (fields.ManyToManyRelation, fields.ReverseRelation)):
+            return list(value)
+        return value
+
+    @classmethod
+    async def from_tortoise_orm(cls, obj: "Model") -> "PydanticModel":
+        # Get fields needed to fetch
+        fetch_fields = _get_fetch_fields(cls, getattr(cls.__config__, "orig_model"))
+        # Fetch fields
+        await obj.fetch_related(*fetch_fields)
+        # Convert to pydantic object
+        values = super().from_orm(obj)
+        return values
+
+
+class PydanticListModel(BaseModel):
+    """ Custom Pydantic BaseModel for List of Tortoise Models """
+
+    @classmethod
+    async def from_queryset(cls, queryset: "QuerySet") -> "PydanticListModel":
+        submodel = getattr(cls.__config__, "submodel")
+        fetch_fields = _get_fetch_fields(submodel, getattr(submodel.__config__, "orig_model"))
+        values = cls(
+            __root__=[submodel.from_orm(e) for e in await queryset.prefetch_related(*fetch_fields)]
+        )
+        return values
